修复数据库连接超时

+++ src/main/java/com/bkit/fatdown/mappers/dao/RolePermissionRelationDao.java
@@ -27,5 +27,5 @@ public interface RolePermissionRelationDao {
      * @param roleId 角色id
      * @return
      */
-    List<TbPermission> getPermissionList(@Param("roleId") Long roleId);
+    List<TbPermission> getPermissionList(@Param("roleId") Integer roleId);
 }
